prediction/requirements.txt
# GRPC Python setup requirements
protobuf>=3.5.0.post1
grpcio-tools
scikit-learn
pandas
